# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateTenantDto": {
        "properties": {
          "civilStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TenantCivilStatus"
              }
            ]
          },
          "cpf": {
            "example": "123.456.789-09",
            "pattern": "^\\d{3}\\.?\\d{3}\\.?\\d{3}-?\\d{2}$",
            "type": "string"
          },
          "email": {
            "example": "locatario@example.com",
            "format": "email",
            "maxLength": 254,
            "type": "string"
          },
          "mobilePhone": {
            "example": "+55 11 99999-9999",
            "maxLength": 20,
            "type": "string"
          },
          "profession": {
            "example": "Engenheiro civil",
            "maxLength": 100,
            "minLength": 2,
            "type": "string"
          },
          "rg": {
            "example": "12.345.678-9",
            "maxLength": 20,
            "minLength": 5,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "cpf",
          "rg",
          "profession",
          "civilStatus",
          "email",
          "mobilePhone"
        ],
        "type": "object"
      },
...
      "PageMetaDto": {
        "properties": {
          "limit": {
            "example": 20,
            "maximum": 100,
            "minimum": 1,
            "type": "number"
          },
          "page": {
            "example": 1,
            "minimum": 1,
            "type": "number"
          },
          "total": {
            "example": 42,
            "minimum": 0,
            "type": "number"
          },
          "totalPages": {
            "example": 3,
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ],
        "type": "object"
      },
...
      "PaginatedTenantResponseDto": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TenantResponseDto"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "ProblemDetailsDto": {
        "properties": {
          "detail": {
            "example": "A requisição contém dados inválidos.",
            "type": "string"
          },
          "errors": {
            "description": "Demais mensagens de validação, quando houver mais de uma.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "instance": {
            "example": "/contracts?page=0",
            "type": "string"
          },
          "requestId": {
            "example": "0f2a9a56-b0d5-4ac5-8f1e-791e9fc92347",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "example": 400,
            "maximum": 599,
            "minimum": 400,
            "type": "number"
          },
          "timestamp": {
            "example": "2026-07-12T18:30:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "title": {
            "example": "BadRequestException",
            "type": "string"
          },
          "type": {
            "example": "https://tenancy-ledger.local/problems/http-400",
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "requestId",
          "timestamp"
        ],
        "type": "object"
      },
...
      "TenantCivilStatus": {
        "enum": [
          "SINGLE",
          "MARRIED",
          "DIVORCED",
          "WIDOWED",
          "STABLE_UNION"
        ],
        "type": "string"
      },
      "TenantResponseDto": {
        "properties": {
          "civilStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TenantCivilStatus"
              }
            ]
          },
          "cpf": {
            "description": "CPF mascarado.",
            "example": "***.***.***-09",
            "type": "string"
          },
          "email": {
            "description": "E-mail mascarado.",
            "example": "l***@example.com",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "mobilePhone": {
            "description": "Telefone mascarado.",
            "example": "(**) *****-9999",
            "type": "string"
          },
          "profession": {
            "example": "Engenheiro civil",
            "type": "string"
          }
        },
        "required": [
          "id",
          "cpf",
          "profession",
          "civilStatus",
          "email",
          "mobilePhone"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "JWT de acesso.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "API para gestão auditável de inquilinos, imóveis, contratos e cobranças.",
    "title": "Tenancy Ledger API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/tenants": {
      "get": {
        "operationId": "TenantController_findAll",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTenantResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Requisição inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Token ausente, inválido ou expirado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Papel sem permissão para a operação.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Limite de requisições excedido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Erro interno inesperado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar locatários",
        "tags": [
          "Locatários"
        ]
      },
      "post": {
        "operationId": "TenantController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Requisição inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Token ausente, inválido ou expirado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Papel sem permissão para a operação.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "CPF ou e-mail já cadastrado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Regra de negócio ou valor de domínio inválido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Limite de requisições excedido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Erro interno inesperado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cadastrar locatário",
        "tags": [
          "Locatários"
        ]
      }
    },
    "/tenants/{id}": {
      "get": {
        "operationId": "TenantController_findById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Requisição inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Token ausente, inválido ou expirado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Papel sem permissão para a operação.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Locatário não encontrado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Limite de requisições excedido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Erro interno inesperado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar locatário",
        "tags": [
          "Locatários"
        ]
      }
    }
  },
  "servers": [],
...
}